block out edit page
